+++ packages/cli/src/commands/add/messages.js
@@ -13,12 +13,6 @@ export function printConfirmationMessage(changeset) {
   const majorReleases = getReleasesOfType("major");
   const minorReleases = getReleasesOfType("minor");
   const patchReleases = getReleasesOfType("patch");
-  const patchDependents = changeset.dependents
-    .filter(dep => dep.type === "patch")
-    .map(dep => dep.name);
-  const majorDependents = changeset.dependents
-    .filter(dep => dep.type === "major")
-    .map(dep => red(dep.name));
 
   if (majorReleases.length > 0)
     logger.log(`${green("[Major]")}\n  ${majorReleases.join(", ")}`);
@@ -26,29 +20,20 @@ export function printConfirmationMessage(changeset) {
     logger.log(`${green("[Minor]")}\n  ${minorReleases.join(", ")}`);
   if (patchReleases.length > 0)
     logger.log(`${green("[Patch]")}\n  ${patchReleases.join(", ")}`);
-  if (patchDependents.length > 0)
-    logger.log(
-      `${green("[Dependents (patch)]")}\n  ${patchDependents.join("\n  ")}`
-    );
-  if (majorDependents.length > 0)
-    logger.log(
-      `${green("[Dependents (major)]")}\n  ${majorDependents.join("\n  ")}`
-    );
 
-  if (changeset.dependents.length > 0) {
-    const message = outdent`
+  const message = outdent`
       ${red("========= NOTE ========")}
-      All dependents that are bumped will be ${red("patch bumped")}.
+      When we need to bump dependents, we will do this as a ${red(
+        "patch bump"
+      )}.
       If any of the above need a higher bump than this, you will need to create a ${red(
         "separate changeset"
-      )} for this
-      Please read the above list ${red(
-        "carefully"
-      )} to make sure you're not missing anything!`;
-    const prettyMessage = boxen(message, {
-      borderStyle: "double",
-      align: "center"
-    });
-    logger.log(prettyMessage);
-  }
+      )}
+      
+      If you want to see what we will currently patch bump, run \`yarn changeset status\``;
+  const prettyMessage = boxen(message, {
+    borderStyle: "double",
+    align: "center"
+  });
+  logger.log(prettyMessage);
 }
